% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hyperparameter optimisation.R
\name{folds_gen}
\alias{folds_gen}
\title{The functions in this file are to be used for hyperparameter optimisation
Hyperparameters: q (k), n(number of components), cint, varmax}
\usage{
folds_gen(df)
}
\description{
The functions in this file are to be used for hyperparameter optimisation
Hyperparameters: q (k), n(number of components), cint, varmax
}
