% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model_functions.R
\name{p_est}
\alias{p_est}
\title{The functionS in this file define the various functions used for running the model}
\usage{
p_est(p, k = NA, rows = NA, interaction_numb = 2, effectsize = 13)
}
\description{
The functionS in this file define the various functions used for running the model
}
