% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Technique.R
\name{plsregfunct}
\alias{plsregfunct}
\title{The functionS in this file define the function used for creating the model}
\usage{
plsregfunct(
  df,
  ncomp = 3,
  interactionterm = T,
  int_term = 3,
  model_perf = c("rsq", "rmse", "bic", NA)
)
}
\description{
The functionS in this file define the function used for creating the model
}
