% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model.R
\name{Prop_mod}
\alias{Prop_mod}
\title{Prepare the model}
\usage{
Prop_mod(
  df_p,
  seeder = 1,
  ncomp = 3,
  i_numb = 2,
  k = 5,
  effectsize = 13,
  model_perf = "None",
  coarse_FS = T,
  Fine_FS = T,
  summary_ci = 0.95,
  strict_anova = T,
  semi_strict = T,
  varmax = 10,
  prediction_type = "rmse",
  predict_method = "reg",
  kmean_type = "Normal"
)
}
\description{
Prepare the model
}
